HAI-2245 Moved /testadata/verified-name to /profiili/verified-name
