+++ services/hanke-service/src/main/resources/email/template/johtoselvitys-valmis.html.mustache
@@ -1,5 +1,5 @@
 <!doctype html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:o="urn:schemas-microsoft-com:office:office">
+<html lang="und" dir="auto" xmlns="http://www.w3.org/1999/xhtml" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:o="urn:schemas-microsoft-com:office:office">
 
 <head>
   <title></title>
@@ -26,12 +26,10 @@
     <![endif]-->
   <style type="text/css">@media only screen and (min-width:480px){.mj-column-per-100{width:100%!important;max-width:100%}}</style>
   <style media="screen and (min-width:480px)">.moz-text-html .mj-column-per-100{width:100%!important;max-width:100%}</style>
-  <style type="text/css"></style>
-  <style type="text/css"></style>
 </head>
 
 <body style="word-spacing:normal">
-  <div>
+  <div lang="und" dir="auto">
     <!--[if mso | IE]><table align="center" border="0" cellpadding="0" cellspacing="0" class="" role="presentation" style="width:600px;" width="600" ><tr><td style="line-height:0px;font-size:0px;mso-line-height-rule:exactly;"><![endif]-->
     <div style="margin:0 auto;max-width:600px">
       <table align="center" border="0" cellpadding="0" cellspacing="0" role="presentation" style="width:100%">
